Pull out editor styles to external file.
